refactor: moves virtual-teams/ team-map related sources to their own folder
