+++ src/main.ts
@@ -4,7 +4,7 @@ import { type Writable } from "node:stream";
 import { parseArgs } from "node:util";
 import generateCodeOwners from "./codeowners/generate.js";
 import generateLabelerYml from "./labeler-yml/generate.js";
-import readTeamMap from "./read-team-map.js";
+import readTeamMap from "./team-map/read.js";
 import readVirtualCodeOwners from "./virtual-code-owners/read.js";
 import { VERSION } from "./version.js";
